% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_condition.R
\name{add_condition}
\alias{add_condition}
\title{Add experimental condition labels}
\usage{
add_condition(data)
}
\arguments{
\item{data}{Simulated dataset.}
}
\value{
Updated dataset.
}
\description{
Adds baseline and cue conditions.
}
