% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missingness.R
\name{add_missingness}
\alias{add_missingness}
\title{Add missing observations}
\usage{
add_missingness(data, missing_rate = 0)
}
\arguments{
\item{data}{Dataset.}

\item{proportion}{Missing proportion.}
}
\value{
Dataset with missing values.
}
\description{
Simulates missing physiological measurements.
}
