% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary_model.R
\name{fit_summary_model}
\alias{fit_summary_model}
\title{Fit summary-statistic model}
\usage{
fit_summary_model(data)
}
\arguments{
\item{data}{Dataset.}
}
\value{
P-value.
}
\description{
Fit summary-statistic model
}
